requirements.txt
# asgiref==3.2.10
astroid==2.5.2
atomicwrites==1.4.0
attrs==20.1.0
beautifulsoup4==4.9.1
colorama==0.4.3
dj-static==0.0.6
Django==3.2
django-bootstrap4==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
iniconfig==1.0.1
isort==5.4.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.8.0
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.7.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
template-remover==0.1.9
toml==0.10.1
wrapt==1.12.1
mock==4.0.2
Pillow==8.0
# factory-boy==3.2.0
gunicorn==20.0.4
psycopg2==2.8.6